+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [Add config for always prompting for nREPL port](https://github.com/BetterThanTomorrow/calva/issues/3162)
+
 ## [2.0.571] - 2026-04-09
 
 - Dev: Fix npm audit: upgrade jest, typescript, and patch transitive deps